# Conflicts:
#	pkg/config/callback_test.go
#	pkg/config/convertxds_test.go
fix conflicts and fix integrate test xds_config_test mistake

+++ etc/script/report.sh
@@ -3,8 +3,13 @@
 set -e
 echo "" > coverage.txt
 
-for d in $(go list ./pkg/... | grep -v scenetest); do
-    go test -coverprofile=profile.out -covermode=atomic $d
+for d in $(go list ./pkg/...); do
+    output="$(go test -v -coverprofile=profile.out -covermode=atomic $d)"
+    echo "$output"
+    fail=$(echo "$output" | grep "\-\-\- FAIL:" | wc -l)
+    if [ $fail -ne 0 ]; then
+	    exit 1
+    fi
     if [ -f profile.out ]; then
         cat profile.out >> coverage.txt
         rm profile.out
